--- lud31617/TAREA7/gestor-tareas-lowdb/.vscode/launch.json
@@ -1,11 +1,0 @@
-{
-  "version": "0.2.0",
-  "configurations": [
-    {
-      "command": "./node_modules/.bin/astro dev",
-      "name": "Development server",
-      "request": "launch",
-      "type": "node-terminal"
-    }
-  ]
-}
