--- lud31617/TAREA7/gestor-tareas-lowdb/db.json
@@ -1,19 +1,0 @@
-{
-  "tareas": [
-    {
-      "id": "mazvv31gf7bcbbu025w",
-      "texto": "sadasdwe",
-      "completada": false
-    },
-    {
-      "id": "mazvv4j9qr09pdfow5",
-      "texto": "asdasdasd",
-      "completada": false
-    },
-    {
-      "id": "mazvv8bswdu0yyswnep",
-      "texto": "sadadssad",
-      "completada": false
-    }
-  ]
-}
