--- lud31617/TAREA7/gestor-tareas-lowdb/src/styles/global.css
@@ -1,1 +1,0 @@
-@import "tailwindcss";
