+++ lud31617/TAREA7ROUTING/gestor-tareas-tanstack/.vscode/extensions.json
@@ -1,0 +1,4 @@
+{
+  "recommendations": ["astro-build.astro-vscode"],
+  "unwantedRecommendations": []
+}
