+++ lud31617/TAREA7ROUTING/gestor-tareas-tanstack/db.json
@@ -1,0 +1,9 @@
+{
+  "tareas": [
+    {
+      "id": "mazvv8bswdu0yyswnep",
+      "texto": "sadadssad",
+      "completada": false
+    }
+  ]
+}
