+++ lud31617/TAREA7ROUTING/gestor-tareas-tanstack/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "tarea-react-lista",
+  "type": "module",
+  "version": "0.0.1",
+  "scripts": {
+    "dev": "astro dev",
+    "build": "astro build",
+    "preview": "astro preview",
+    "astro": "astro"
+  },
+  "dependencies": {
+    "@astrojs/react": "^4.2.7",
+    "@tailwindcss/vite": "^4.1.6",
+    "@tanstack/react-query": "^5.76.2",
+    "@types/react": "^19.1.4",
+    "@types/react-dom": "^19.1.5",
+    "astro": "^5.7.13",
+    "lowdb": "^7.0.1",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "react-hot-toast": "^2.5.2",
+    "tailwindcss": "^4.1.6",
+    "zustand": "^5.0.5"
+  }
+}
