+++ lud31617/TAREA7ROUTING/gestor-tareas-tanstack/src/pages/NotFound.tsx
@@ -1,0 +1,3 @@
+export default function NotFound() {
+  return <h1 className="text-3xl text-center mt-10">404 – Página no encontrada</h1>;
+}
