+++ lud31617/TAREA7ROUTING/gestor-tareas-tanstack/src/pages/SettingsPage.tsx
@@ -1,0 +1,3 @@
+export default function SettingsPage() {
+  return <h2 className="text-2xl">Settings (pendiente de implementar)</h2>;
+}
