+++ lud31617/TAREA7ROUTING/gestor-tareas-tanstack/src/pages/index.astro
@@ -1,0 +1,14 @@
+---
+import TaskApp from '../components/TaskApp.tsx';
+import '../styles/global.css';
+---
+
+<html lang="es">
+  <head>
+    <meta charset="UTF-8" />
+    <title>Lista de Tareas</title>
+  </head>
+  <body class="bg-gradient-to-r from-blue-100 to-indigo-100 min-h-screen py-10">
+    <TaskApp client:load />
+  </body>
+</html>
