+++ lud31617/TAREA7ROUTING/gestor-tareas-tanstack/src/styles/global.css
@@ -1,0 +1,1 @@
+@import "tailwindcss";
